feat: Add `NullifierWitness` type
